requirements.txt
speech_recognition==3.10.0
pyttsx3==2.90
googletrans==4.0.0rc1
keyboard==0.13.5
requests==2.31.0
python-dotenv==1.0.0
PyAudio==0.2.11